requirements.txt
flask==2.1.3
pandas==1.4.4
plotly_express==0.4.1
python-dotenv==0.21.0
requests==2.28.1
matplotlib==3.5.2
seaborn==0.12.0
scikit-learn==1.1.3
termcolor==2.1.0